Add more dummy functionality, fix some error handling, and organize views
