+++ backend/coreapp/m2c_wrapper.py
@@ -1,7 +1,6 @@
 import contextlib
 import io
 import logging
-from typing import Optional
 
 from mips_to_c.src.main import parse_flags
 from mips_to_c.src.main import run
@@ -21,7 +20,6 @@ class M2CWrapper:
             # Create temp asm file
             asm_path = sandbox.path / "asm.s"
             asm_path.write_text(asm)
-
 
             if context:
                 # Create temp context file
